ARMS-PROJECT/requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-migrate
flask-script
pymysql
flask-jwt-extended
requests
discord.py
python-dotenv
pytest
pytest-cov
Werkzeug==0.16.0